Move sampling method
